Order BLL

+++ WindowSeller/WindowSellerWASM/Shared/Persistance/IGenericRepository.cs
@@ -8,11 +8,11 @@ namespace WindowSellerWASM.Shared.Persistance
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> GetAsync(int id);
+        Task<T> GetAsync(long id);
         Task<IReadOnlyList<T>> GetAllAsync();
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(T entity);
-        Task<bool> DoesExistAsync(int id);
+        Task<bool> DoesExistAsync(long id);
     }
 }
